move unired-tag config to its own file
