+++ unired-tag/index.js
@@ -1,25 +1,9 @@
 const {
-  constants: { ERRORS, TIMEFRAMES },
+  constants: { ERRORS },
 } = require('@notify-watcher/core');
+const config = require('./config');
 
 const UNIRED_URL = 'https://www.unired.cl';
-
-const config = {
-  name: 'unired-tag-total',
-  displayName: 'Unired Tag Total',
-  description: 'Watch your tag ballot using Unired',
-  auth: true,
-  libs: ['puppeteer'],
-  timeframe: {
-    type: TIMEFRAMES.minute,
-  },
-  notificationTypes: {
-    updatedBallot: {
-      key: 'updatedBallot',
-      description: 'Your Tag ballot changed.',
-    },
-  },
-};
 
 async function waitAndClick(page, selector) {
   await page.waitForSelector(selector);
